planning-service/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.2
django-click==2.3.0
django-health-check==3.17.0
django-extensions==3.2.1
django-utils==0.0.2
djangorestframework==3.14.0
envparse==0.2.0
gunicorn==20.1.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.4
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.12
../events

confluent_kafka==1.9.2
fastavro==1.6.1
jsonschema==4.17.0
cloudevents==1.7.0